porcentaje de aprobados, desaprobados y total pruebas
