+++ src/app/components/statics-widgets/model-tested-widget/model-tested-widget.component.html
@@ -1,0 +1,4 @@
+<p-chart
+    type="bar"
+    [data]="data"
+    [options]="chartOptions"></p-chart>
