+++ src/app/features/statics/statics.component.html
@@ -25,22 +25,36 @@
         title="Modelos que más aprobaron - Top {{ top }}"
         class="col col-6 p-2">
         <app-model-approved-widget
-            [tags]="metricsDataMap.meterModelApproved"
-            [top]="top"></app-model-approved-widget>
+            [approvedTags]="metricsDataMap.meterModelApproved"
+            [failedTags]="metricsDataMap.meterModelFailed"
+            [top]="10">
+        </app-model-approved-widget>
     </app-static-widget>
 
     <app-static-widget
         title="Modelos que más desaprobaron - Top {{ top }}"
         class="col col-6 p-2">
         <app-model-failed-widget
-            [tags]="metricsDataMap.meterModelFailed"
-            [top]="top"></app-model-failed-widget>
+            [approvedTags]="metricsDataMap.meterModelApproved"
+            [failedTags]="metricsDataMap.meterModelFailed"
+            [top]="10">
+        </app-model-failed-widget>
     </app-static-widget>
 
     <app-static-widget
         title="Utilización de los puestos"
         class="col col-6 p-2">
         <app-stand-used-widget [tags]="metricsDataMap.standUsed"></app-stand-used-widget>
+    </app-static-widget>
+
+    <app-static-widget
+        title="Modelos más testeados - Top {{ top }}"
+        class="col col-6 p-2">
+        <app-model-tested-widget
+            [approvedTags]="metricsDataMap.meterModelApproved"
+            [failedTags]="metricsDataMap.meterModelFailed"
+            [top]="10">
+        </app-model-tested-widget>
     </app-static-widget>
 </div>
